refactor(cli): give headroom to graph, task, discover at 300-line cap
The crates/convergio-cli audit follow-up (T828d03c) flagged three
files within three lines of the 300-line cap. None had room left
to extend safely, so split orthogonal concerns into siblings:

 271): move `ForTaskHints`,
  `append_query_param`, and `encode_query_value` into a new
  `graph_query` sibling, gaining unit tests for the percent
  encoder.
 280): move `parse_github_pr_url` into a
  new `task_pr_url` sibling, gaining unit tests for the
  canonical URL, the non-pull rejection, and query/fragment
  stripping.
 178): move `render_human` and
  `render_plain` (plus the shared `parse_dt` helper) into a
  new `discover_render` sibling.

Also drop a dead `let _ = file; // file glob would need a
directory  skip for now` placeholder inscan
`docs_generators::gen_adr_index` flagged by the same audit
(LOW finding on docs_generators.rs:156).

No behavior change for any of `cvg graph`, `cvg task`,
`cvg discover`, or `cvg docs regenerate`. The full
convergio-cli unit-test suite passes (113 tests, all green).

Tracks: T828d03c7-4019-4c60-9a13-88f2404e1f9d
