+++ crates/convergio-cli/src/commands/mod.rs
@@ -25,6 +25,7 @@ pub mod crdt;
 pub mod dash;
 pub mod demo;
 pub mod discover;
+mod discover_render;
 pub mod dispatch;
 pub mod docs;
 mod docs_generators;
@@ -43,6 +44,7 @@ pub(crate) mod fleet_duplicates;
 pub(crate) mod fleet_patterns;
 pub mod gates;
 pub mod graph;
+mod graph_query;
 mod graph_render;
 pub mod health;
 pub mod mcp;
@@ -67,6 +69,7 @@ pub mod status;
 mod status_render;
 pub mod task;
 mod task_complete;
+mod task_pr_url;
 mod task_render;
 pub mod task_templates;
 pub mod update;
